Add toggleable 'applied' button on watch list page
Still need logic to do this on main list. Currently the 'applied' property only exists on the items in the watch list and not the items in the main job array. I think I could maybe add a cross check with the items in the watched list just after the main job array populates. then add the property accordingly.